Move analytical queries from constants.py to separate files
When debugging or developing, it is often necessary to run the CH2(++)
queries manually or share them with others. The query definitions in
`constants.py` made this more painful than necessary due to their python
string format, and also blew up `constants.py`. Separating out each of
the analytical query sets into dedicated files makes it easier to read
the queries and work with them manually when necessary.

+ Move analytical queries to files in
  ch2driver/pytpcc/analytical_queries
+ Remove analytical query dictionaries from constants.py
+ Add `loadAnalyticalQueriesFromFile` method to `AbstractDriver` class
+ Update `nestcollectionsdriver.py` and `mongodbdriver.py` to load
  analytical queries from files
+ Format and tidy `abstractdriver.py`
+ Update ch2driver/pytpcc/README.md

Change-Id: I906f7d6cbab1321627aa641db7602e29a70cd35b
Reviewed-on: https://review.couchbase.org/c/ch2/+/235998
Tested-by: Daniel Nagy <daniel.nagy@couchbase.com>
Reviewed-by: Daniel Nagy <daniel.nagy@couchbase.com>